DDP-5758 - remove FormInstanceDaoTest (move all tests to ActivityInstanceServiceTest) as class FormInstanceDao was removed too
